Factor out update source tests (#626)
