+++ tests/Functional/Application/UpdateSourceTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Integration;
+namespace App\Tests\Functional\Application;
 
 use App\Tests\Application\AbstractUpdateSourceTest;
